requirements.txt
Django==4.2.4
stripe==7.6.0
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==20.1.0